Restore attached email parsing when $TreatAttachedEmailAsFiles is not set
For cases where RT parses emails that come in as attachments to
other email, restore the previous parsing of additional headers.
This will show email addresses from attached email as suggestions on
the People page and on reply. This is useful to more easily set up
roles on a ticket in cases where someone forwards email into RT
that was intended for another person or queue.

The $TreatAttachedEmailAsFiles option now provides an alternate
behavior for email attachments that doesn't try to parse attached
email. When this is set, get email addresses only from the parent
attachment.