Improve UI of unread messages notification
When a user has opted to receive notifications of unread messages,
a panel at the top of the ticket display page gives the user the
option to mark messages as read and also jump to the first unread
message. The previous UI used links within the text making it
difficult to see. UI updated with buttons instead. Previous
variable, "UpdateStatus" was changed to "UnreadMessages" to make
it clear that it refers to tickets with unread messages.
Corresponding tests also updated accordingly.

Fixes: I#30082
Fixes: I#32584